move DBInstance to separate file, prepare for its thread-safety
