+++ libosmscout-client-qt/src/osmscout/DBThread.cpp
@@ -29,72 +29,6 @@
 #include <osmscout/TextSearchIndex.h>
 #endif
 
-QBreaker::QBreaker()
-  : osmscout::Breaker(),
-    aborted(false)
-{
-}
-
-void QBreaker::Break()
-{
-  QMutexLocker locker(&mutex);
-
-  aborted=true;
-}
-
-bool QBreaker::IsAborted() const
-{
-  QMutexLocker locker(&mutex);
-
-  return aborted;
-}
-
-void QBreaker::Reset()
-{
-  QMutexLocker locker(&mutex);
-
-  aborted=false;
-}
-
-StyleError::StyleError(QString msg){
-    QRegExp rx("(\\d+),(\\d+) (Symbol|Error|Warning|Exception):(.*)");
-    if(rx.exactMatch(msg)){
-        line = rx.cap(1).toInt();
-        column = rx.cap(2).toInt();
-        if(rx.cap(3) == "Symbol"){
-            type = Symbol;
-        } else if(rx.cap(3) == "Error"){
-            type = Error;
-        } else if(rx.cap(3) == "Warning"){
-            type = Warning;
-        } else {
-            type = Exception;
-        }
-        text = rx.cap(4);
-    }
-}
-
-QString StyleError::GetTypeName() const
-{
-    switch(type){
-    case Symbol:
-        return QString("symbol");
-        break;
-    case Error:
-        return QString("error");
-        break;
-    case Warning:
-        return QString("warning");
-        break;
-    case Exception:
-        return QString("exception");
-        break;
-    default:
-      return QString("???");
-    }
-}
-
-
 DBThread::DBThread(QStringList databaseLookupDirs,
                    QString iconDirectory,
                    SettingsRef settings)
@@ -487,104 +421,6 @@ void DBThread::LoadStyle(QString stylesheetFilename,
   InvalidateVisualCache();
   emit stylesheetFilenameChanged();
   emit Redraw();
-}
-
-bool DBInstance::LoadStyle(QString stylesheetFilename,
-                           std::unordered_map<std::string,bool> stylesheetFlags,
-                           QList<StyleError> &errors)
-{
-
-
-  if (!database->IsOpen()) {
-    return false;
-  }
-
-  osmscout::TypeConfigRef typeConfig=database->GetTypeConfig();
-
-  if (!typeConfig) {
-    return false;
-  }
-
-  // new map style may require more data types. when tile is marked as "completed"
-  // such data types are never loaded into these tiles
-  // so we mark them as "incomplete" to make sure that all types for new stylesheet are loaded
-  mapService->InvalidateTileCache();
-  osmscout::StyleConfigRef newStyleConfig=std::make_shared<osmscout::StyleConfig>(typeConfig);
-
-  for (const auto& flag : stylesheetFlags) {
-    newStyleConfig->AddFlag(flag.first,flag.second);
-  }
-
-  if (newStyleConfig->Load(stylesheetFilename.toLocal8Bit().data())) {
-    // Tear down
-    if (painter!=NULL){
-      delete painter;
-      painter=NULL;
-    }
-
-    // Recreate
-    styleConfig=newStyleConfig;
-    painter=new osmscout::MapPainterQt(styleConfig);
-  }
-  else {
-    std::list<std::string> errorsStrings=newStyleConfig->GetErrors();
-
-    for(const auto& errorString : errorsStrings) {
-      StyleError err(QString::fromStdString(errorString));
-      qWarning() << "Style error:" << err.GetDescription();
-      errors.append(err);
-    }
-
-    styleConfig=NULL;
-
-    return false;
-  }
-
-  return true;
-}
-
-
-bool DBInstance::AssureRouter(osmscout::Vehicle /*vehicle*/,
-                              osmscout::RouterParameter routerParameter)
-{
-  if (!database->IsOpen()) {
-    return false;
-  }
-
-  if (!router/* ||
-      (router && router->GetVehicle()!=vehicle)*/) {
-    if (router) {
-      if (router->IsOpen()) {
-        router->Close();
-      }
-      router=NULL;
-    }
-
-    router=std::make_shared<osmscout::RoutingService>(database,
-                                                      routerParameter,
-                                                      osmscout::RoutingService::DEFAULT_FILENAME_BASE);
-
-    if (!router->Open()) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-void DBInstance::close()
-{
-  if (router && router->IsOpen()) {
-    router->Close();
-  }
-
-  if (callbackId){
-    mapService->DeregisterTileStateCallback(callbackId);
-  }
-  callbackId = 0;
-  if (database->IsOpen()) {
-    database->Close();
-  }
 }
 
 bool DBThread::GetObjectDetails(DBInstanceRef db,
